+++ vllm/model_executor/layers/fused_moe/__init__.py
@@ -87,6 +87,9 @@ if HAS_TRITON:
         cutlass_moe_w4a8_fp8,
     )
     from vllm.model_executor.layers.fused_moe.deep_gemm_moe import DeepGemmExperts
+    from vllm.model_executor.layers.fused_moe.experts.rocm_aiter_moe import (
+        AiterExperts,
+    )
     from vllm.model_executor.layers.fused_moe.fused_batched_moe import (
         BatchedTritonExperts,
     )
@@ -95,9 +98,6 @@ if HAS_TRITON:
         TritonWNA16Experts,
         fused_experts,
         get_config_file_name,
-    )
-    from vllm.model_executor.layers.fused_moe.rocm_aiter_fused_moe import (
-        AiterExperts,
     )
     from vllm.model_executor.layers.fused_moe.router.fused_topk_router import (
         fused_topk,
